requirements.txt
vosk==0.3.45
sounddevice==0.4.6
webrtcvad==2.0.10
openai>=0.27.0
requests>=2.28.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
pydub>=0.25.1
sqlalchemy>=1.4.0
pgvector>=0.7.2
elevenlabs>=0.2.0
pytest>=7.0.0
